feat: mark unread when scroll switch
